rename folder, add OccurTestData as CSV
