+++ src/DB/seeders/6-20210924092735-occurrence.js
@@ -5,13 +5,14 @@ const results = [];
 
 module.exports = {
     up: async (queryInterface, Sequelize) => {
-        fs.createReadStream('../DB/JsonData/CompleteOccurrences.csv')
+        fs.createReadStream('../DB/Data/TestDataOccurrence.csv')
             .pipe(csv({ separator: ';' }))
-            .on('data', (data) => results.push({
-                workId: data.File,
-                term: data.OccId,
-                scientificName: data.Term,
-                wordsFrame: data['1000-Words-Frame'],
+            .on('data',  (data) => results.push(
+              {
+                workId: data.workId,
+                term: data.Term,
+                scientificName: data['Scientific Name'],
+                wordsFrame: data.WordsFrame,
                 sentence: data.Sentence,
                 createdAt: new Date(),
                 updatedAt: new Date()
@@ -20,7 +21,6 @@ module.exports = {
                 console.log(results);
                 return queryInterface.bulkInsert('Occurrences', results)
             });
-
     },
 
     down: async (queryInterface, Sequelize) => {
